Branch : feature_resolve_issue_float
- Handle crash for Float price
  Convert int to float
  ISSUE : When order submit then order details price come in int (i.e. 6.5 is giving 7)

+++ VegeApp/src/com/sabziatdoor/adapter/ProductAdapter.java
@@ -34,7 +34,7 @@ public class ProductAdapter extends BaseAdapter {
 	
 	
 	public interface ITotalCount {
-		public void getTotal(int updateValue,int prize);
+		public void getTotal(int updateValue,float prize);
 	}
 	
 	ITotalCount iTotalCount = null;
@@ -100,7 +100,7 @@ public class ProductAdapter extends BaseAdapter {
 			     		productQtyInCart=0;
 			     		mDatabaseHelper.deleteMyCartItem(rowItem.getProductId());
 			     	} else {
-			     		int productsubTotal=productQtyInCart*Integer.parseInt(rowItem.getProductMainPrice());
+			     		Float productsubTotal=productQtyInCart*Float.parseFloat(rowItem.getProductMainPrice());
 				     	ContentValues cartValue=new ContentValues();
 				     	cartValue.put(VegAppColumn.CART_PRODUCT_QTY,productQtyInCart);
 				     	cartValue.put(VegAppColumn.CART_PRODUCT_SUB_TOTAL,productsubTotal);
@@ -109,7 +109,7 @@ public class ProductAdapter extends BaseAdapter {
 			     	mDatabaseHelper.close();
 			     	holder.txtViewQty.setText(""+productQtyInCart);
 					rowItem.setProductQty(productQtyInCart);
-					iTotalCount.getTotal(-1,Integer.parseInt(rowItem.getProductMainPrice()));
+					iTotalCount.getTotal(-1,Float.parseFloat(rowItem.getProductMainPrice()));
 				}
 			}
 		});
@@ -123,14 +123,14 @@ public class ProductAdapter extends BaseAdapter {
 				int productQtyInCart=mDatabaseHelper.getCartProductQty(rowItem.getProductId());
 				if(productQtyInCart!=-1){
 					productQtyInCart+=1;
-			     	int productsubTotal=productQtyInCart*Integer.parseInt(rowItem.getProductMainPrice());
+					Float productsubTotal=productQtyInCart*Float.parseFloat(rowItem.getProductMainPrice());
 			     	ContentValues cartValue=new ContentValues();
 			     	cartValue.put(VegAppColumn.CART_PRODUCT_QTY,productQtyInCart);
 			     	cartValue.put(VegAppColumn.CART_PRODUCT_SUB_TOTAL,productsubTotal);
 			     	mDatabaseHelper.updateMyCart(rowItem.getProductId(),cartValue);
 				} else {
 			     	productQtyInCart=1;
-			     	int productsubTotal=productQtyInCart*Integer.parseInt(rowItem.getProductMainPrice());
+			     	Float productsubTotal=productQtyInCart*Float.parseFloat(rowItem.getProductMainPrice());
 
 			     	ContentValues cartValue=new ContentValues();
 			     	
@@ -155,7 +155,7 @@ public class ProductAdapter extends BaseAdapter {
 				holder.txtViewQty.setText(""+productQtyInCart);
 				rowItem.setProductQty(productQtyInCart);				
 				mDatabaseHelper.close();
-				iTotalCount.getTotal(1,Integer.parseInt(rowItem.getProductMainPrice()));
+				iTotalCount.getTotal(1,Float.parseFloat(rowItem.getProductMainPrice()));
 			}
 		});
         holder.txtViewProductName.setText(rowItem.getProductName());
